python/requirements.txt
fastapi
uvicorn
ortools
google-generativeai
pydantic
